refactor(api): create one package per handler (#2507)
