+++ api/handlers/node/node_test.go
@@ -1,4 +1,4 @@
-package handlers
+package node
 
 import (
 	"context"
