+++ api/handlers/validators/validators_test.go
@@ -1,4 +1,4 @@
-package handlers
+package validators
 
 import (
 	"encoding/json"
